Added Board, IMPORTANT: you will need boost
Using the boost library for the board.
The board consists of a single dimensional vector of nrCols * nrRows
boost::optional< TileOnBoard >. (The boost::optional makes it easy to
work with empty spaces on the board.)
You will need boost (preferably boost 1.52.0) and install it in the
right folder to use this.